+++ wiselib.testing/algorithms/neighbor_discovery/neighbor_discovery.h
@@ -785,12 +785,12 @@ namespace wiselib
 			return NULL;
 		}
 		// --------------------------------------------------------------------
-		int get_status()
+		uint8_t get_status()
 		{
 			return status;
 		}
 		// --------------------------------------------------------------------
-		void set_status( int _st )
+		void set_status( uint8_t _st )
 		{
 			status = _st;
 		}
